Reorganize folder structure for components
