+++ client/src/components/Dashboard/PlayerProfiles.js
@@ -6,8 +6,8 @@ import ProfileStatTable from './ProfileStatTable'
 import IconButton from '@material-ui/core/IconButton'
 import { TableChart, Close } from '@material-ui/icons'
 import { connect } from 'react-redux'
-import { removePlayerList, openPlayerModal } from '../actions/statActions'
-import { yearFormatter } from '../helper/columnLabels'
+import { removePlayerList, openPlayerModal } from '../../actions/statActions'
+import { yearFormatter } from '../../helper/columnLabels'
 const mobileWidth = '425px'
 const tabletWidth = '800px'
